// app-angular/src/api/se/model/arcScriptEvalResponse.ts
/**
 * Storage Explorer Embedded RESTful API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {ArcScriptColumnDescriptor} from './arcScriptColumnDescriptor';
import {ArcScriptEvalError} from './arcScriptEvalError';


export interface ArcScriptEvalResponse {
    columns?: Array<ArcScriptColumnDescriptor>;
    resultSet: Array<object>;
    err?: ArcScriptEvalError;
}