// app-angular/src/api/se/model/storageEntryType.ts
/**
 * Storage Explorer Embedded RESTful API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum StorageEntryType {
    LIST = 'LIST',
    MAP = 'MAP',
    SEQUENCE = 'SEQUENCE',
    OBJECT = 'OBJECT'
}